Fix Array That Stores Reading Status
Previously when trying to put the read status in a book in the correct spot
it would always get pushed to the end of the array. The read status now matches
the index that the book is stored in. Now the Array just needs the be updated
whenever a book is added or deleted.