Add a 3rd Party Method of Toggling Read Status
Built a custom button that could toggle the read status of a book
without directly clicking on the read button itself. This will be
useful b/c it makes me one step closwer towards having the myLibrary
for each loop being able to apply the status of a book based on the index
of the readStorage array. The goal is to fix the issue where The read
status of the books are reset to default after a new book is added