Almost Stop Check Marks Dissapearing When Add Book
Check boxes now stay on the proper index when new book is added.

All that needs fixing it how the readStorage Array update.
If I fully delete the books from the display without unchecking the read
button, newly added books will be added with the checkbox already active.
The readStorage needs to properly upadte in real time so that everything
is in-sync.